requirements.txt
Django==4.2.2
django-filter==23.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7
celery==5.1.2
redis==4.3.4